requirements.txt
flickrapi
httplib2
Pillow
tweepy
urllib3
flask
bs4